engine/requirements.txt
openai>=1.12.0
elevenlabs>=1.0.0
yt-dlp>=2024.1.0
chromadb>=0.4.22
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.1
python-telegram-bot>=21.0
langchain>=0.1.5
langchain-openai>=0.0.5
pypdf>=4.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
